Simplify Ice.rb
