+++ ruby/ruby/Ice/Exception.rb
@@ -1,0 +1,22 @@
+#
+# Copyright (c) ZeroC, Inc. All rights reserved.
+#
+
+module Ice
+
+    class Exception < ::StandardError
+    end
+
+    class LocalException < Exception
+    end
+
+    class UserException < Exception
+        def ice_id
+            to_s
+        end
+
+        def inspect
+            return ::Ice::__stringifyException(self)
+        end
+    end
+end
