+++ ruby/ruby/Ice/IdentitySpaceship.rb
@@ -1,0 +1,18 @@
+#
+# Copyright (c) ZeroC, Inc. All rights reserved.
+#
+
+require_relative 'Identity.rb'
+
+module Ice
+    class Identity
+        def <=>(other)
+            n = self.name <=> other.name
+            if n == 0
+                return self.category <=> other.category
+            else
+                return n
+            end
+        end
+    end
+end
