+++ ruby/ruby/Ice/InitializationData.rb
@@ -1,0 +1,14 @@
+#
+# Copyright (c) ZeroC, Inc. All rights reserved.
+#
+
+module Ice
+    class InitializationData
+        def initialize(properties=nil, logger=nil)
+            @properties = properties
+            @logger = logger
+        end
+
+        attr_accessor :properties, :logger
+    end
+end
